Flask/requirements.txt
streamlit
flask
pandas
mysql-connector-python
scikit-learn
requests
openpyxl
